+++ lib/inferno/dsl/fhir_resource_validation.rb
@@ -30,6 +30,13 @@ module Inferno
         klass.extend ClassMethods
       end
 
+      # Find a particular profile StructureDefinition and the IG it belongs to.
+      # Looks through a runnable's parents up to the suite to find a validator with a particular name,
+      # then finds the profile by looking through its defined igs.
+      def find_ig_and_profile(profile_url, validator_name)
+        self.class.find_ig_and_profile(profile_url, validator_name)
+      end
+
       class Validator
         attr_reader :requirements
         attr_accessor :session_id, :name, :test_suite_id
@@ -402,6 +409,20 @@ module Inferno
           current_validators << new_validator
 
           fhir_validators[name] = current_validators
+        end
+
+        # @private
+        def find_ig_and_profile(profile_url, validator_name)
+          validator = find_validator(validator_name)
+          if validator.is_a? Inferno::DSL::FHIRResourceValidation::Validator
+            validator.igs.each do |ig_id|
+              ig = Inferno::Repositories::IGs.new.find_or_load(ig_id)
+              profile = ig.profile_by_url(profile_url)
+              return ig, profile if profile
+            end
+          end
+
+          raise "Unable to find profile #{profile_url} in any IG defined for validator #{validator_name}"
         end
       end
     end
